# DOCS/contracts/openapi.json
{
...
  "openapi": "3.0.1",
  "info": {
    "title": "ConfluxPay API",
    "description": "ConfluxPay backend API â€” multi-tenant payment platform. All requests require a tenant API key carried in the Authorization header.",
    "contact": {
      "name": "ConfluxPay Engineering",
      "email": "engineering@conflux.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://conflux.com/license"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8089",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/merchant/kyc": {
      "post": {
        "tags": [
          "Merchant Onboarding"
        ],
        "operationId": "submitKyc",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KycSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultMerchantOnboardingDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/mfa/disable": {
      "post": {
        "tags": [
          "MFA"
        ],
        "summary": "Disable MFA (requires password)",
        "operationId": "disable",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaDisableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users/{id}/unblock": {
      "post": {
        "tags": [
          "Admin - Merchants"
        ],
        "operationId": "unblockUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/block": {
      "post": {
        "tags": [
          "Admin - Merchants"
        ],
        "operationId": "blockUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/merchants/{id}/verify": {
      "post": {
        "tags": [
          "Admin - Merchants"
        ],
        "operationId": "verifyMerchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/merchants/{id}/reject": {
      "post": {
        "tags": [
          "Admin - Merchants"
        ],
        "operationId": "rejectMerchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMerchantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/merchants": {
      "get": {
        "tags": [
          "Admin - Merchants"
        ],
        "operationId": "listMerchants",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "REGISTERED",
                "PENDING_VERIFICATION",
                "VERIFIED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListMerchantSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResultMeta": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "RESOURCE_NOT_FOUND",
              "RESOURCE_ALREADY_EXISTS",
              "INVALID_OPERATION_STATE",
              "INSUFFICIENT_FUNDS",
              "IDEMPOTENCY_CONFLICT",
              "MFS_ADAPTER_FAILURE",
              "VENDOR_DOWN",
              "QUOTA_EXCEEDED",
              "RISK_REJECTED",
              "TOKEN_EXPIRED",
              "INTERNAL_ERROR"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResultMerchantOnboardingDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MerchantOnboardingDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "MerchantOnboardingDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "identifier": {
            "type": "string"
          },
          "identifierType": {
            "type": "string",
            "enum": [
              "PHONE",
              "EMAIL",
              "USERNAME"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "BLOCKED"
            ]
          },
          "fullName": {
            "type": "string"
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "REGISTERED",
              "PENDING_VERIFICATION",
              "VERIFIED",
              "REJECTED"
            ]
          }
        }
      },
      "KycSubmissionRequest": {
        "required": [
          "nidBackUrl",
          "nidFrontUrl",
          "tradeLicenseUrl"
        ],
        "type": "object",
        "properties": {
          "nidFrontUrl": {
            "type": "string"
          },
          "nidBackUrl": {
            "type": "string"
          },
          "tradeLicenseUrl": {
            "type": "string"
          },
          "tinUrl": {
            "type": "string"
          }
        }
      },
...
      "ApiResultVoid": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
...
      "MfaDisableRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "BlockUserRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "RejectMerchantRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResultListMerchantSummaryDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MerchantSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "MerchantSummaryDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "merchantProfileId": {
            "type": "string",
            "format": "uuid"
          },
          "identifier": {
            "type": "string"
          },
          "identifierType": {
            "type": "string",
            "enum": [
              "PHONE",
              "EMAIL",
              "USERNAME"
            ]
          },
          "fullName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "BLOCKED"
            ]
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "REGISTERED",
              "PENDING_VERIFICATION",
              "VERIFIED",
              "REJECTED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "description": "Tenant API key. Send as `Authorization: Bearer <api-key>`. Keys are issued per tenant in Phase 1; rotation invalidates previous keys immediately.",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  }
}